+++ Mesh_3/include/CGAL/odt_optimize_mesh_3.h
@@ -83,7 +83,7 @@ odt_optimize_mesh_3_impl(C3T3& c3t3,
     max_iteration_number = 1000;
   
   // Launch optimization
-  return opt(max_iteration_number);
+  return opt(static_cast<int>(max_iteration_number));
 }
 
 }  // end namespace CGAL
